刘鸥 吴世豪 吕东

+++ study-api/src/main/java/com/kgc/study/service/loService.java
@@ -1,5 +1,7 @@
 package com.kgc.study.service;
 
+
+
 import com.kgc.study.bean.EduUser;
 import com.kgc.study.bean.MyPageInfo;
 
@@ -8,7 +10,7 @@ import java.util.List;
 public interface loService {
                                      //用户名                   手机号               身份证              性别                 是否可用                     是否是好学生             入学方式
     MyPageInfo<EduUser> selectEduUser(
-            String eduUserName,String eduUserPhone,String eduUserIdCard,Byte eduUserGender,String eduUserIsAvalible,Integer eduUserIsGoodStudent,Byte eduUserEnrollmentMethod,
+            String eduUserName,String eduUserPhone,String eduUserIdCard,Integer eduUserGender,String eduUserIsAvalible,Integer eduUserIsGoodStudent,Integer eduUserEnrollmentMethod,
             Integer pageNum, Integer pageSize);
 
     int insertEduUserList(List<EduUser> eduUsers);
